nidd_api/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
# defense.src.detector / features need these; omitting them causes ImportError → HTTP 500 on /api/predict
numpy>=1.24.0
scipy>=1.10.0
scikit-learn==1.8.0
twilio>=9.0.0
